Use u64 consistently for EXPECTED GROUP SIZE
It's unclear why Reduce originally took this as a usize, but it must
be a 64-bit value, and we use it for determining bucketing that has
nothing to do with the word-size of the machine mz is running on.

This doesn't change anything in practice, except that on 32-bit
machines, I think the code wouldn't have even built, since `cast_from`
lacks an implementation for usize from u64.

+++ src/expr/src/relation/mod.rs
@@ -221,7 +221,7 @@ pub enum MirRelationExpr {
         monotonic: bool,
         /// User hint: expected number of values per group key. Used to optimize physical rendering.
         #[serde(default)]
-        expected_group_size: Option<usize>,
+        expected_group_size: Option<u64>,
     },
     /// Groups and orders within each group, limiting output.
     ///
@@ -1234,7 +1234,7 @@ impl MirRelationExpr {
         self,
         group_key: Vec<usize>,
         aggregates: Vec<AggregateExpr>,
-        expected_group_size: Option<usize>,
+        expected_group_size: Option<u64>,
     ) -> Self {
         MirRelationExpr::Reduce {
             input: Box::new(self),
